+++ services/storage_daemon/ipc/test/storage_daemon_test.cpp
@@ -30,7 +30,7 @@
 #include "test/common/help_utils.h"
 #include "user/user_manager.h"
 #include "utils/file_utils.h"
-
+#include "ipc/enum_daemon.h"
 namespace OHOS {
 namespace StorageDaemon {
 using namespace testing::ext;
@@ -104,8 +104,7 @@ HWTEST_F(StorageDaemonTest, Storage_Manager_StorageDaemonTest_PrepareUserDirs_00
     auto bRet = StorageTest::StorageTestUtils::CreateFile(filePath);
     EXPECT_TRUE(bRet) << "check the file create";
 
-    int32_t flags = static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL1) |
-                    static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL2);
+    int32_t flags = IStorageDaemonEnum::CRYPTO_FLAG_EL1 | IStorageDaemonEnum::CRYPTO_FLAG_EL2;
     int32_t ret = storageDaemon_->PrepareUserDirs(StorageTest::USER_ID1, flags);
     EXPECT_TRUE(ret == E_PREPARE_DIR) << "the path is not dir";
 
@@ -124,7 +123,7 @@ HWTEST_F(StorageDaemonTest, Storage_Manager_StorageDaemonTest_PrepareUserDirs_00
 
     ASSERT_TRUE(storageDaemon_ != nullptr);
 
-    int32_t flags = static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL1);
+    int32_t flags = IStorageDaemonEnum::CRYPTO_FLAG_EL1;
     auto ret = storageDaemon_->PrepareUserDirs(StorageTest::USER_ID2, flags);
     EXPECT_TRUE(ret == E_OK) << "the flags is incorrect";
 
@@ -143,8 +142,7 @@ HWTEST_F(StorageDaemonTest, Storage_Manager_StorageDaemonTest_PrepareUserDirs_00
 
     ASSERT_TRUE(storageDaemon_ != nullptr);
 
-    int32_t flags = static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL1) |
-                    static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL2);
+    int32_t flags = IStorageDaemonEnum::CRYPTO_FLAG_EL1 | IStorageDaemonEnum::CRYPTO_FLAG_EL2;
     auto ret = storageDaemon_->PrepareUserDirs(StorageTest::USER_ID3, flags);
     EXPECT_TRUE(ret == E_OK);
 
@@ -182,8 +180,7 @@ HWTEST_F(StorageDaemonTest, Storage_Manager_StorageDaemonTest_StartUser_002, Tes
 
     ASSERT_TRUE(storageDaemon_ != nullptr);
 
-    int32_t flags = static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL1) |
-                    static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL2);
+    int32_t flags = IStorageDaemonEnum::CRYPTO_FLAG_EL1 | IStorageDaemonEnum::CRYPTO_FLAG_EL2;
     auto ret = storageDaemon_->PrepareUserDirs(StorageTest::USER_ID5, flags);
     EXPECT_TRUE(ret == E_OK) << "create user dirs error";
 
@@ -207,8 +204,7 @@ HWTEST_F(StorageDaemonTest, Storage_Manager_StorageDaemonTest_DestroyUserDirs_00
 
     ASSERT_TRUE(storageDaemon_ != nullptr);
 
-    int32_t flags = static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL1) |
-                    static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL2);
+    int32_t flags = IStorageDaemonEnum::CRYPTO_FLAG_EL1 | IStorageDaemonEnum::CRYPTO_FLAG_EL2;
     auto ret = storageDaemon_->PrepareUserDirs(StorageTest::USER_ID3, flags);
     EXPECT_TRUE(ret == E_OK);
 
@@ -229,8 +225,7 @@ HWTEST_F(StorageDaemonTest, Storage_Manager_StorageDaemonTest_StopUser_001, Test
     GTEST_LOG_(INFO) << "Storage_Manager_StorageDaemonTest_StopUser_001 start";
 
     ASSERT_TRUE(storageDaemon_ != nullptr);
-    int32_t flags = static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL1) |
-                    static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL2);
+    int32_t flags = IStorageDaemonEnum::CRYPTO_FLAG_EL1 | IStorageDaemonEnum::CRYPTO_FLAG_EL2;
     auto ret = storageDaemon_->PrepareUserDirs(StorageTest::USER_ID1, flags);
     EXPECT_TRUE(ret == E_OK) << "create user dirs error";
 
@@ -257,8 +252,7 @@ HWTEST_F(StorageDaemonTest, Storage_Manager_StorageDaemonTest_StopUser_002, Test
 
     ASSERT_TRUE(storageDaemon_ != nullptr);
 
-    int32_t flags = static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL1) |
-                    static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL1);
+    int32_t flags = IStorageDaemonEnum::CRYPTO_FLAG_EL1 | IStorageDaemonEnum::CRYPTO_FLAG_EL1;
     auto ret = storageDaemon_->PrepareUserDirs(StorageTest::USER_ID4, flags);
     EXPECT_TRUE(ret == E_OK) << "create user dirs error";
 
@@ -281,8 +275,7 @@ HWTEST_F(StorageDaemonTest, Storage_Manager_StorageDaemonTest_StopUser_003, Test
 
     ASSERT_TRUE(storageDaemon_ != nullptr);
 
-    int32_t flags = static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL1) |
-                    static_cast<uint32_t>(IStorageDaemonEnum::CRYPTO_FLAG_EL2);
+    int32_t flags = IStorageDaemonEnum::CRYPTO_FLAG_EL1 | IStorageDaemonEnum::CRYPTO_FLAG_EL2;
     auto ret = storageDaemon_->PrepareUserDirs(StorageTest::USER_ID3, flags);
     EXPECT_TRUE(ret == E_OK) << "PrepareUserDirs error";
     ret = storageDaemon_->StartUser(StorageTest::USER_ID3);
